# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.CreatePlayerPayload:
    properties:
      bio:
        type: string
      hourly_rate_cents:
        type: integer
      main_game_id:
        type: integer
      nickname:
        type: string
      user_id:
        type: integer
      verification_status:
        type: string
    required:
    - user_id
    - verification_status
    type: object
...
  admin.SkillTagsBody:
    properties:
      tags:
        items:
          type: string
        type: array
    required:
    - tags
    type: object
...
  admin.UpdatePlayerPayload:
    properties:
      bio:
        type: string
      hourly_rate_cents:
        type: integer
      main_game_id:
        type: integer
      nickname:
        type: string
      verification_status:
        type: string
    required:
    - verification_status
    type: object
...
info:
  contact: {}
  description: GameLink 平台 API，包含健康检查、认证与管理端能力。
  title: GameLink API
  version: 0.3.0
paths:
...
  /admin/players:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出玩家资料
      tags:
      - Admin/Players
    post:
      consumes:
      - application/json
      parameters:
      - description: 玩家信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreatePlayerPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新建玩家资料
      tags:
      - Admin/Players
  /admin/players/{id}:
    delete:
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除玩家资料
      tags:
      - Admin/Players
    get:
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取玩家资料
      tags:
      - Admin/Players
    put:
      consumes:
      - application/json
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      - description: 玩家信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.UpdatePlayerPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新玩家资料
      tags:
      - Admin/Players
  /admin/players/{id}/games:
    put:
      consumes:
      - application/json
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{main_game_id}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            format: int64
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新玩家主游戏
      tags:
      - Admin/Players
  /admin/players/{id}/logs:
    get:
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 动作过滤
        enum:
        - create
        - update
        - delete
        in: query
        name: action
        type: string
      - description: 操作者用户ID
        in: query
        name: actor_user_id
        type: integer
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      - description: 导出格式
        enum:
        - csv
        in: query
        name: export
        type: string
      - description: 导出列（逗号分隔）
        in: query
        name: fields
        type: string
      - description: 列头语言
        enum:
        - en
        - zh
        in: query
        name: header_lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取玩家操作日志
      tags:
      - Admin/Players
...
  /admin/players/{id}/skill-tags:
    put:
      consumes:
      - application/json
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签集合
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.SkillTagsBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新玩家技能标签
      tags:
      - Admin/Players
  /admin/players/{id}/verification:
    put:
      consumes:
      - application/json
      parameters:
      - description: 玩家ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{verification_status}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新玩家认证状态
      tags:
      - Admin/Players
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"